requirements.txt
dj-database-url==0.5.0
Django==2.2.6
gunicorn==19.9.0
Pillow==6.2.0
python-decouple==3.1
pytz==2019.2
sqlparse==0.3.0
whitenoise==4.1.4
certifi==2019.9.11
chardet==3.0.4
future==0.17.1
idna==2.8
oauthlib==3.1.0
python-decouple==3.1
# Add python-twitter
python-twitter==3.5
requests==2.22.0
requests-oauthlib==1.2.0
urllib3==1.25.6
